Refactor linear/mod.rs into submodules for maintainability
Split the largest lint-flagged file (805 lines) into smaller modules:
- mod.rs: 172 lines (config, helpers, re-exports)
- forge_impl.rs: 499 lines (Forge trait implementation)
- link.rs: 142 lines (link flow)

All files now comply with the 500-line limit.
